+++ lib/ui/page/login/view.dart
@@ -19,13 +19,13 @@ import 'package:animated_size_and_fade/animated_size_and_fade.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:messenger/ui/widget/outlined_rounded_button.dart';
 
-import '/domain/model/user.dart';
+import '/domain/model/my_user.dart';
 import '/l10n/l10n.dart';
 import '/themes.dart';
 import '/ui/page/home/page/chat/widget/chat_item.dart';
 import '/ui/widget/modal_popup.dart';
+import '/ui/widget/outlined_rounded_button.dart';
 import '/ui/widget/primary_button.dart';
 import '/ui/widget/svg/svg.dart';
 import '/ui/widget/text_field.dart';
@@ -42,6 +42,7 @@ class LoginView extends StatelessWidget {
   const LoginView({
     super.key,
     this.initial = LoginViewStage.signUp,
+    this.myUser,
     this.onSuccess,
   });
 
@@ -53,15 +54,19 @@ class LoginView extends StatelessWidget {
   /// If not specified, the [RouteLinks.home] redirect is invoked.
   final void Function({bool? signedUp})? onSuccess;
 
+  /// [MyUser], whose data should be prefilled in the fields.
+  final MyUser? myUser;
+
   /// Displays a [LoginView] wrapped in a [ModalPopup].
   static Future<T?> show<T>(
     BuildContext context, {
     LoginViewStage initial = LoginViewStage.signUp,
+    MyUser? myUser,
     void Function({bool? signedUp})? onSuccess,
   }) {
     return ModalPopup.show(
       context: context,
-      child: LoginView(initial: initial, onSuccess: onSuccess),
+      child: LoginView(initial: initial, myUser: myUser, onSuccess: onSuccess),
     );
   }
 
@@ -74,6 +79,7 @@ class LoginView extends StatelessWidget {
       init: LoginController(
         Get.find(),
         initial: initial,
+        myUser: myUser,
         onSuccess: onSuccess,
       ),
       builder: (LoginController c) {
@@ -490,7 +496,7 @@ class LoginView extends StatelessWidget {
                   final bool enabled = !c.login.isEmpty.value &&
                       !c.password.isEmpty.value &&
                       c.signInTimeout.value == 0 &&
-                      c.authStatus.value.isEmpty;
+                      !c.authStatus.value.isLoading;
 
                   return PrimaryButton(
                     key: const Key('LoginButton'),
@@ -540,21 +546,38 @@ class LoginView extends StatelessWidget {
               header = ModalPopupHeader(text: 'label_sign_in'.l10n);
 
               children = [
-                PrimaryButton(
-                  title: 'btn_sign_in'.l10n,
+                Obx(() {
+                  return SignButton(
+                    key: const Key('StartButton'),
+                    icon: const SvgIcon(SvgIcons.guest),
+                    padding: const EdgeInsets.only(left: 4),
+                    onPressed: c.authStatus.value.isEmpty ? c.register : () {},
+                    title: 'btn_guest'.l10n,
+                  );
+                }),
+                const SizedBox(height: 15),
+                SignButton(
+                  key: const Key('RegisterButton'),
+                  icon: const SvgIcon(SvgIcons.register),
+                  padding: const EdgeInsets.only(left: 3),
+                  onPressed: () {
+                    c.returnTo = c.stage.value;
+                    c.stage.value = LoginViewStage.signUp;
+                  },
+                  title: 'btn_sign_up'.l10n,
+                ),
+                const SizedBox(height: 15),
+                SignButton(
+                  key: const Key('SignInButton'),
+                  icon: const SvgIcon(SvgIcons.enter),
+                  padding: const EdgeInsets.only(left: 4),
                   onPressed: () {
                     c.returnTo = c.stage.value;
                     c.stage.value = LoginViewStage.signIn;
                   },
+                  title: 'btn_sign_in'.l10n,
                 ),
-                const SizedBox(height: 25 / 2),
-                Obx(() {
-                  return PrimaryButton(
-                    title: 'btn_guest'.l10n,
-                    onPressed: c.authStatus.value.isEmpty ? c.register : () {},
-                  );
-                }),
-                const SizedBox(height: 16),
+                const SizedBox(height: 15),
               ];
               break;
           }